Implement authentication and login flow, and wire it into app
This is a rough implementation of authentication and login flow. It is good
enough for dev purposes and to prevent arbitrary access attempts.

This is also the first "working" version so there are lots of miscellaneous
changes to make it work. In particular, there were bugs with the connection
logic and the html templating, and various bugs/typos.