Implement a basic API endpoint to fetch tagged data
This is really dumb - it fetches all the data for the signed in user. It is
good enough for now though to serve mock data to the front end so the graphs
can be worked on.

Also, created a script to populate some dummy data. This script is highly
coupled to the state of the local db.