Set up lightweight test framework for Javascript
Existing frameworks that I looked into were all too heavy handed for my
taste, given the early state of the project. The only dependency I brought
in was "assert". Most of the work was caused by my needing to transpile
and bundle the Javascirpt in a way that mirrors jspm.